Исправления старого файла, а не нового. добавлены тесты
