+++ index.php
@@ -1,4 +1,4 @@
 <?php
-require_once("vendor/autoload.php");
 require_once("test.php");
 echo isTaskCorrect();
+//Добавил тесты
